organized some components category
