+++ client/app.js
@@ -1,12 +1,9 @@
 import React from 'react';
-
-import {Navbar, Landing} from './components';
 import Routes from './routes';
 
 const App = () => {
   return (
     <div>
-      {/*<Navbar />*/}
       <Routes />
     </div>
   );
