+++ client/components/AccommodationComponents/LocationSearchInput.js
