--- client/components/Expenses.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import Sidebar from './Sidebar';
-
-const Expenses = () => {
-  return (
-    <div>
-      <Sidebar />
-      <div style={{marginLeft: '100px'}}>
-        <h1>Expenses go here</h1>
-      </div>
-    </div>
-  );
-};
-
-export default Expenses;
