+++ client/components/TravelComponents/TravelForm.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import axios from 'axios';
-import {fetchSelected} from '../store/trip';
+import {fetchSelected} from '../../store/trip';
 import {connect} from 'react-redux';
 import DayPickerInput from 'react-day-picker/DayPickerInput';
 import 'react-day-picker/lib/style.css';
