+++ public/style.css
@@ -16,127 +16,6 @@
   width: 100%;
 }
 
-.activities-map-container {
-  margin-top: 50px;
-  background: white;
-  border-radius: 5px;
-}
-
-.map-container {
-  width: 100%;
-  height: 460px;
-  border-radius: 5px;
-}
-
-.map-footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 10px;
-  margin-top: 40px;
-  font-family: 'Istok Web', sans-serif;
-}
-.map-footer button {
-  height: 30px;
-  width: 80px;
-  border: #c60057;
-  border-radius: 3px;
-  background: #c60057;
-  color: white;
-  font-size: 18px;
-  font-family: 'Istok Web', sans-serif;
-}
-
-.activity-modal {
-  z-index: 100;
-  margin-top: -180px;
-  left: 400px;
-  display: flex;
-  justify-content: center;
-  top: 300px;
-  position: absolute;
-  border: 1px solid rgb(185, 185, 185);
-  border-radius: 15px;
-  width: 350px;
-  background: white;
-  text-align: center;
-  height: 450px;
-}
-
-.activity-modal-content {
-  margin: 0 auto;
-}
-
-.modal-form {
-  margin: 0 auto;
-}
-
-.modal-form input {
-  border: 1px solid #c60057;
-  height: 25px;
-  margin-bottom: 10px;
-  width: 230px;
-}
-.modal-form button {
-  margin-bottom: 10px;
-  background: #c60057;
-  color: white;
-  font-size: 20px;
-  border: none;
-  border-radius: 3px;
-  width: 85px;
-  height: 60px;
-}
-
-.activity-list {
-  text-align: right;
-  width: 300px;
-  margin: 0px auto;
-  background: white;
-  border-radius: 5px;
-  margin-top: 50px;
-}
-
-.activity-list-header {
-  color: white;
-  background: #53aad9;
-  border-radius: 3px 3px 0px 0px;
-  height: 30px;
-  font-size: 1.3em;
-  margin: 0;
-  font-family: 'Istok Web', sans-serif;
-}
-
-.activity-list-item {
-  margin: 3px;
-  color: white;
-  height: 35px;
-  line-height: 35px;
-  display: flex;
-  justify-content: flex-end;
-  width: 296px;
-  margin-right: 1px;
-}
-
-.activity-list-item-name {
-  margin-right: 3px;
-  font-size: 80%;
-  border-radius: 3px;
-  background-color: #ad20ff;
-  width: 260px;
-  text-align: center;
-}
-
-.activity-list button {
-  color: white;
-  height: 100%;
-  border: none;
-  background: red;
-  text-align: center;
-  width: 30px;
-  cursor: pointer;
-}
-
 .landing-header {
   text-align: right;
   margin: 0 auto;
@@ -1094,7 +973,6 @@ nav a {
 }
 
 .accommodation-container,
-.activities-container,
 .participants-container {
   text-align: center;
   width: 400px;
@@ -1106,8 +984,7 @@ nav a {
   height: 250px;
 }
 
-.accommodation-container-list,
-.activities-container-list {
+.accommodation-container-list {
   text-align: center;
   margin: 3px;
 }
@@ -1131,7 +1008,6 @@ nav a {
 }
 
 .accommodation-container-header,
-.activities-container-header,
 .participants-container-header {
   color: white;
   background-color: #53aad9;
@@ -1145,7 +1021,6 @@ nav a {
 }
 
 .accommodation-container-header span,
-.activities-container-header span,
 .participants-container-header span {
   margin-left: 8px;
   font-family: 'Istok Web', sans-serif;
@@ -1373,9 +1248,199 @@ button:disabled {
   align-items: center;
 }
 
-.activities-container-list span,
 .accommodation-container-list span {
   color: #c60057;
   font-family: 'Istok Web';
   font-size: 1.3em;
+}
+
+.activities-container {
+  text-align: center;
+  width: 400px;
+  margin-top: 20px;
+  background-color: white;
+  font-family: 'Istok Web', sans-serif;
+  box-shadow: -7px 7px 6px -2px rgba(0, 0, 0, 0.75);
+  border-radius: 3px;
+  height: 250px;
+}
+
+.activities-container-header {
+  color: white;
+  background-color: #53aad9;
+  height: 40px;
+  display: flex;
+  justify-content: space-between;
+  font-size: 1.3em;
+  line-height: 40px;
+  border-top-left-radius: 3px;
+  border-top-right-radius: 3px;
+}
+
+.activities-container-header span {
+  margin-left: 8px;
+  font-family: 'Istok Web', sans-serif;
+}
+
+.activities-container-list {
+  text-align: center;
+  margin: 3px;
+}
+
+.activities-container-list span {
+  color: #c60057;
+  font-family: 'Istok Web';
+  font-size: 1.3em;
+}
+
+.activity-form-add {
+  width: 180px;
+  height: 45px;
+  border: 2px solid #c60057;
+  display: flex;
+  justify-content: space-evenly;
+  margin: 0 auto;
+  border-radius: 15px;
+  align-items: center;
+  cursor: pointer;
+  width: 280px;
+}
+
+.activity-form-add-plus {
+  color: white;
+  background-color: #c60057;
+  border: 2px solid #c60057;
+  border-radius: 50px;
+  width: 35px;
+  height: 35px;
+  font-size: 3em;
+  line-height: 40px;
+}
+
+.activity-form-add-text {
+  color: #c60057;
+  font-weight: bold;
+}
+
+.activities-map-container {
+  margin-top: 50px;
+  background: white;
+  border-radius: 5px;
+}
+
+.map-container {
+  width: 100%;
+  height: 460px;
+  border-radius: 5px;
+}
+
+.map-footer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 10px;
+  margin-top: 40px;
+  font-family: 'Istok Web', sans-serif;
+}
+
+.map-footer button {
+  height: 30px;
+  width: 80px;
+  border: #c60057;
+  border-radius: 3px;
+  background: #c60057;
+  color: white;
+  font-size: 18px;
+  font-family: 'Istok Web', sans-serif;
+}
+
+.activity-modal {
+  z-index: 100;
+  margin-top: -180px;
+  left: 400px;
+  display: flex;
+  justify-content: center;
+  top: 300px;
+  position: absolute;
+  border: 1px solid rgb(185, 185, 185);
+  border-radius: 15px;
+  width: 350px;
+  background: white;
+  text-align: center;
+  height: 450px;
+}
+
+.activity-modal-content {
+  margin: 0 auto;
+}
+
+.modal-form {
+  margin: 0 auto;
+}
+
+.modal-form input {
+  border: 1px solid #c60057;
+  height: 25px;
+  margin-bottom: 10px;
+  width: 230px;
+}
+
+.modal-form button {
+  margin-bottom: 10px;
+  background: #c60057;
+  color: white;
+  font-size: 20px;
+  border: none;
+  border-radius: 3px;
+  width: 85px;
+  height: 60px;
+}
+
+.activity-list {
+  text-align: right;
+  width: 300px;
+  margin: 0px auto;
+  background: white;
+  border-radius: 5px;
+  margin-top: 50px;
+}
+
+.activity-list-header {
+  color: white;
+  background: #53aad9;
+  border-radius: 3px 3px 0px 0px;
+  height: 30px;
+  font-size: 1.3em;
+  margin: 0;
+  font-family: 'Istok Web', sans-serif;
+}
+
+.activity-list-item {
+  margin: 3px;
+  color: white;
+  height: 35px;
+  line-height: 35px;
+  display: flex;
+  justify-content: flex-end;
+  width: 296px;
+  margin-right: 1px;
+}
+
+.activity-list-item-name {
+  margin-right: 3px;
+  font-size: 80%;
+  border-radius: 3px;
+  background-color: #ad20ff;
+  width: 260px;
+  text-align: center;
+}
+
+.activity-list button {
+  color: white;
+  height: 100%;
+  border: none;
+  background: red;
+  text-align: center;
+  width: 30px;
+  cursor: pointer;
 }
